pyproject.toml
[project]
name = "audio-rag"
version = "0.1.0"
description = "Production-grade, config-driven Audio RAG system with multilingual support"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "edge-tts>=7.2.7",
    "faster-whisper>=1.2.1",
    "ffmpeg-python>=0.2.0",
    "nvidia-cudnn-cu12>=9.10.2.21",
    "piper-tts>=1.2.0",
    "pyannote-audio>=4.0.3",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.14.3",
    "sentence-transformers>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "pre-commit>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/audio_rag"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/audio_rag"]
branch = true
